feat: add honeypots
This enables the creation of honeypots.
Each honeypot is a specific instance that we can connect an attacker to.
Moreover, each honeypot is assigned to exactly one router, since the honeypots
represent specific instances, like containers or hardware appliances, that are
located at a specific site.
Dynamic honeypots will likely be created by the routers themselves, while static
honeypots will be created via the UI.

We first thought about modelling the honeypot route as `/routers/:routerId/honeypots/:honeypotId`.
However, we came across the question, what happens when requesting
`/routers/nonExistingRouterId/honeypots/existingHoneypotId`. As the honeypots have a globally
unique ID, the id of the router is not really important here. However, we would still have to
check whether this specific router exists, as otherwise returning the specific honeypot would
be odd. More specifically, this applies to the FindById, Modify, and Delete operations, in all
of which the routerId should have been checked.
Therefore, we decided to flatten the structure. However, to keep the logical relation to the
router, creating a new endpoint is still done via `POST /routers/:routerId/honeypots`.
We also thought about adding redirects from `/routers/:routerId/honeypots/:honeypotId` to
`/honeypot/:honeypotId` but found that to be an odd overhead.

For more discussion on that, refer to <https://stackoverflow.com/q/20951419>.