fix: make response properties required and add count of array elements if missing
Since we always want to return the properties, it makes sense to mark
them as required. The count for arrays should also always be present.
Additionally, if a property is not marked as required, oapi-codegen
generates pointers instead of values in the api client objects and we
would always have to perform nil checks, which is unnecessary and ugly.