feat: connect objects with teams
This connects objects such as endpoints with teams.
When creating a new object that belongs to a team, the route is now
/team/:teamId/object.
However, objects are still queried using their global route like
/object/:objectId.

Users are a little different, as they can be part of multiple teams. Hence,
users are added to a team via 'PUT /teams/:teamId/users/:userId' and removed
the same way using the DELETE operation.

Objects can be transferred to other teams using their specific /transfer route.