feat: dont load permissions for all objects in array
This adds the permissions array at the same level as the
object arrays, so that we return the permissions for the
whole array instead of only single objects.
As we return these permissions on the team routes, this
should be fine, since the permissions for an object are
determined based on the role inside the team.
This also allows us to convey that users can create new
objects in the array, when the permissions array contains
the create value.