feat: add management of WireGuard information
This adds that bees and beehives now report their WireGuard public key when
registering with the beekeeper. In return, they receive their WireGuard IP
which they should use to configure themselves.

Moreover, beehives and bees now receive their peers in addition to the forwarding
rules so that they can enable the WireGuard connection.